to_utf8 becomes a common function
